Removed components file and put all items into 1. Was a little unnecessary
