+++ app/src/main/java/com/brandoncano/capacitorcalculator/ui/screens/learn/CommonCodesScreen.kt
@@ -2,9 +2,13 @@ package com.brandoncano.capacitorcalculator.ui.screens.learn
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material3.Scaffold
@@ -13,13 +17,18 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.brandoncano.capacitorcalculator.R
+import com.brandoncano.capacitorcalculator.data.CapacitorCodeConversions
 import com.brandoncano.capacitorcalculator.ui.theme.CapacitorCalculatorTheme
 import com.brandoncano.sharedcomponents.composables.AppDivider
 import com.brandoncano.sharedcomponents.composables.AppScreenPreviews
 import com.brandoncano.sharedcomponents.composables.AppTopAppBar
+import com.brandoncano.sharedcomponents.text.onSurfaceVariant
 import com.brandoncano.sharedcomponents.text.textStyleBody
+import com.brandoncano.sharedcomponents.text.textStyleHeadline
+import com.brandoncano.sharedcomponents.text.textStyleSubhead
 
 @Composable
 fun CommonCodesScreen(
@@ -42,6 +51,14 @@ fun CommonCodesScreen(
 @Composable
 private fun CommonCodesScreenContent(paddingValues: PaddingValues) {
     val sidePadding = dimensionResource(R.dimen.app_side_padding)
+    val codes = CapacitorCodeConversions.entries
+    val labels = listOf(
+        stringResource(id = R.string.common_codes_code),
+        stringResource(id = R.string.common_codes_pf),
+        stringResource(id = R.string.common_codes_nf),
+        stringResource(id = R.string.common_codes_uf),
+    )
+
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -53,9 +70,48 @@ private fun CommonCodesScreenContent(paddingValues: PaddingValues) {
             modifier = Modifier.padding(top = 24.dp, bottom = 24.dp),
             style = textStyleBody()
         )
-        ChartRowLabels()
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 8.dp)
+        ) {
+            labels.forEach {
+                Text(
+                    text = it,
+                    modifier = Modifier.weight(1f),
+                    style = textStyleHeadline(),
+                    textAlign = TextAlign.Center,
+                )
+            }
+        }
         AppDivider(modifier = Modifier.padding(horizontal = 0.dp))
-        ChartTable()
+
+        LazyColumn {
+            itemsIndexed(codes) { index, code ->
+                ChartTableRow(code)
+                if (codes.size - 1 != index) {
+                    AppDivider(modifier = Modifier.padding(horizontal = 0.dp))
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun ChartTableRow(row: CapacitorCodeConversions) {
+    Row(
+        modifier = Modifier
+            .padding(vertical = 12.dp)
+            .fillMaxWidth()
+    ) {
+        row.asList().forEach { value ->
+            Text(
+                text = value,
+                modifier = Modifier.weight(1f),
+                style = textStyleSubhead().onSurfaceVariant(),
+                textAlign = TextAlign.Center,
+            )
+        }
     }
 }
